+++ packages/benchmark-js/test/benchmark.ts
@@ -3,6 +3,7 @@ import { Benchmark, BenchmarkGroup, BenchmarkLoggerLevel } from '../src';
 Benchmark.loggerLevel = BenchmarkLoggerLevel.Debug;
 
 const benchmark = new BenchmarkGroup();
+benchmark.setEmptyTest();
 benchmark.add('Benchmark-1', () => {});
 benchmark.add('Benchmark-2', () => {});
 benchmark.add('Benchmark-3', () => {});