+++ packages/benchmark-js/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Run pre-benchmarking double times for optimization.
 - Improve the accuracy of the minimum duration time.
+- Add method `BenchmarkGroup.setEmptyTest` to get the maximum ops as a base line.
 
 ### Bug Fixes